pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Paulescu <plabartabajo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jupyter = "^1.0.0"
llama-index = "^0.10.29"
ragas = "^0.1.7"
langchain-openai = "^0.1.3"
datasets = "^2.18.0"
sentence-transformers = "^2.7.0"
qdrant-client = "^1.8.2"
python-dotenv = "^1.0.1"
tqdm = "^4.66.2"
fire = "^0.6.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
extend-select = ["I"]